pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evb-test"
version = "0.1.0"
description = "Lightweight Python test framework for remote server/device testing"
requires-python = ">=3.10"
dependencies = [
    "paramiko>=3.4.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.8",
    "ruff>=0.3.0",
]

[project.scripts]
evb-test = "evbtest.cli:main"

[tool.setuptools.packages.find]
include = ["evbtest*"]

[tool.ruff]
line-length = 100
target-version = "py310"
